Added permission for Edit task and Delete Task

+++ app/routes.py
@@ -29,14 +29,18 @@ def register_post():
 		username = form.username.data
 		password = form.password.data
 
-		user = Users(name=name, email=email, username=username)
-		user.set_password(password)
-		db.session.add(user)
-		db.session.commit()
-
-		flash("You are now registered and can login", "success")
-
-		return redirect(url_for('login_get'))
+		try:
+			user = Users(name=name, email=email, username=username)
+			user.set_password(password)
+			db.session.add(user)
+			db.session.commit()
+		except Exception:
+			print("in databaseerror exception block")
+			flash("Could not register", "danger")
+			return redirect(url_for('register_get'))
+		else:
+			flash("You are now registered and can login", "success")
+			return redirect(url_for('login_get'))
 
 	flash("Check the values", "danger")
 	return render_template("register.html", form=form)
@@ -113,13 +117,17 @@ def addTask_post():
 		status = False
 
 		# Insert tasks into database
-		task = TaskList(title=title, body=body, status=status, user_id=current_user.id)
-		db.session.add(task)
-		db.session.commit()
+		try:
+			task = TaskList(title=title, body=body, status=status, user_id=current_user.id)
+			db.session.add(task)
+			db.session.commit()
+		except Exception:
+			flash("Problem Adding the task", "danger")
+			return redirect(url_for('userhome'))
+		else:
+			flash("Task added successfully", "success")
+			return redirect(url_for('userhome'))
 		
-		flash("Task added successfully", "success")
-		return redirect(url_for('userhome'))
-
 	return render_template('addTask.html', form=form)
 
 # Edit task
@@ -141,29 +149,40 @@ def editTask_post(id):
 	# Get data
 	task = TaskList.query.filter_by(id=id).first()
 	
+	if task is not None and task.is_own_task(task.user_id):
+		return render_template('editTask.html', task=task)
+
 	# Get Task Details
 	task.title = request.form['title']
 	task.body = request.form['body']
 	task.status = False
-
-	db.session.add(task)
-	db.session.commit()
-
-	flash("Task Updated successfully", "success")
-
-	return redirect(url_for('userhome'))
+	try:
+		db.session.add(task)
+		db.session.commit()
+	except Exception:
+		flash("Couldn't update the task", "danger")
+		return redirect(url_for('userhome'))
+	else:
+		flash("Task Updated successfully", "success")
+		return redirect(url_for('userhome'))
 
 # Delete Task
 @app.route('/deleteTask/<string:id>', methods=['POST'])
 @login_required
 def deleteTask(id):
-	task = TaskList.query.filter_by(id=id).first()
-	db.session.delete(task)
-	db.session.commit()
-
-	flash("Deleted succesfully ", "danger")
-	return redirect(url_for('userhome'))
-
+	try:
+		task = TaskList.query.filter_by(id=id).first()
+		if task is not None and task.is_own_task(task.user_id):
+			return render_template('editTask.html', task=task)
+		db.session.delete(task)
+		db.session.commit()
+	except Exception:
+		flash("Could't Delete the Task", "danger")
+		return redirect(url_for('userhome'))
+	else:
+		flash("Deleted succesfully ", "success")
+		return redirect(url_for('userhome'))
+	
 # Status Of Task
 @app.route('/statusOfTask/<string:status>/<string:id>', methods = ['POST'])
 @login_required
@@ -174,14 +193,19 @@ def statusOfTask(status,id):
 	else:
 		status = False
 	
-	task = TaskList.query.filter_by(id=id).first()
-	task.status = status
-	db.session.add(task)
-	db.session.commit()
-
-	flash("Status changed succesfully ", "success")
-	return redirect(url_for('userhome'))	
-
+	try:
+		task = TaskList.query.filter_by(id=id).first()
+		if task is not None and task.is_own_task(task.user_id):
+			return render_template('editTask.html', task=task)
+		task.status = status
+		db.session.add(task)
+		db.session.commit()
+	except Exception:
+		flash("Status culdn't be changed", "danger")
+		return redirect(url_for('userhome'))	
+	else:
+		flash("Status changed succesfully ", "success")
+		return redirect(url_for('userhome'))	
 
 # Logout
 @app.route('/logout')
